==> R/00_retrieve_EMP_biom.R <==


rm(list = ls())
gc()

options(timeout = 3600)


ftp_site <- "http://ftp.microbio.me/emp/release1/"


output_folder <- "data-raw"

dir.create(output_folder, showWarnings = FALSE)


# subset 2k -----------------------------

download.file(
    
    url = paste0(
        ftp_site, "mapping_files/", "emp_qiime_mapping_subset_2k.tsv"
    ),
    
    method = "curl",
    
    destfile = paste0(output_folder, "/emp_qiime_mapping_subset_2k.tsv")
    
)


download.file(
    
    url = paste0(
        ftp_site, "otu_tables/deblur/", "emp_deblur_90bp.subset_2k.biom"
    ),
    
    method = "curl",
    
    destfile = paste0(output_folder, "/emp_deblur_90bp.subset_2k.biom")
    
)


# subset 5k --------------------------------------

download.file(
    
    url = paste0(
        ftp_site, "mapping_files/", "emp_qiime_mapping_subset_5k.tsv"
    ),
    
    method = "curl",
    
    destfile = paste0(output_folder, "/emp_qiime_mapping_subset_5k.tsv")
    
)

download.file(
    
    url = paste0(
        ftp_site, "otu_tables/deblur/", "emp_deblur_90bp.subset_5k.biom"
    ),
    
    method = "curl",
    
    destfile = paste0(output_folder, "/emp_deblur_90bp.subset_5k.biom")
    
)

# subset 10k ------------------------------------

download.file(
    
    url = paste0(
        ftp_site, "mapping_files/", "emp_qiime_mapping_subset_10k.tsv"
    ),
    
    method = "curl",
    
    destfile = paste0(output_folder, "/emp_qiime_mapping_subset_10k.tsv")
    
)

download.file(
    
    url = paste0(
        ftp_site, "otu_tables/deblur/", "emp_deblur_90bp.subset_10k.biom"
    ),
    
    method = "curl",
    
    destfile = paste0(output_folder, "/emp_deblur_90bp.subset_10k.biom")
    
)

# subset emp release 1 ----------------------------

download.file(
    
    url = paste0(
        ftp_site, "mapping_files/", "emp_qiime_mapping_qc_filtered.tsv"
    ),
    
    method = "curl",
    
    destfile = paste0(output_folder, "/emp_qiime_mapping_qc_filtered.tsv")
    
)

download.file(
    
    url = paste0(
        ftp_site, "otu_tables/deblur/", "emp_deblur_90bp.qc_filtered.biom"
    ),
    
    method = "curl",
    
    destfile = paste0(output_folder, "/emp_deblur_90bp.qc_filtered.biom")
    
)
